% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Using data from FARS  it creates a map of an USA state which contains positions
(longitude and latitude) of accidents from a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer number giving the id of the state}

\item{year}{An integer number giving the year}
}
\value{
Draw a map (of \code{state.num}) with the positions
(longitude and latitude) of accidents from \code{year}.

If the \code{state.num} is invalid stop execution and print a message.

If there aren't accidents in \code{year} a message will be print.
}
\description{
fars_data.zip (FARS dataset) represents the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
This data are from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.
}
\examples{
fars_map_state(24, 2014)
fars_map_state(1, 2015)
\dontrun{
fars_map_state(7,2013)
}

}
