% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Reads a file from disk in a dataframe.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
If there is a file on disk then the output will be a dataframe.
If the file doesn't exist  stops execution and print an error message.
}
\description{
Reads a file from disk in a dataframe.
}
\examples{

fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
\dontrun{
fars_read("accident_2.csv.bz2")
}

}
