% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{From a vector of years creates a list of dataframes with information from
FARS dataset.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers, each element of the vector representing
one year.}
}
\value{
a list with dataframes, each dataframe has just the columns MONTH and year.
In the case that the function don't find information about a year
(from vector \code{years}) in FARS a warning message will be returned.
}
\description{
fars_data.zip (FARS dataset) represents the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
This data are from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.
}
\examples{
fars_read_years(c(2013, 2014, 2))
fars_read_years(c(2013, 2014))


}
