% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{From a vector of years creates a summary of number of accidents grouped by month
and year. Those information are from FARS dataset.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers - each value represents a year}
}
\value{
a dataframe in wide format contains a summary of number of accidents on
each combination year and month. In the case that the function don't find
information about a year (from vector \code{years}) in FARS, a warning message it
will be returned before to return dataframe.
}
\description{
fars_data.zip (FARS dataset) represents the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
This data are from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.
}
\examples{
fars_summarize_years(c(2013, 2014, 2))
fars_summarize_years(c(2013, 2014))

}
