% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Using year as input creates a string consistent with a files names from
FARS dataset.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer number giving the year.}
}
\value{
a string consistent with a files names from FARS dataset.
}
\description{
fars_data.zip (FARS dataset) represents the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.
This data are from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.
}
\examples{
make_filename(2013)
make_filename(2014)
}
